requirements.txt
tweety-ns==1.0.9.5
aiogram==3.7.0